Services optimization

+++ microservice.user/src/main/java/com/bookstore/microservice/user/services/UserService.java
@@ -1,15 +1,14 @@
 package com.bookstore.microservice.user.services;
 
 import com.bookstore.microservice.user.dto.UserDTO;
-import com.bookstore.microservice.user.domain.User;
 
 import java.util.List;
 
 public interface UserService {
     List<UserDTO> getAllUsers();
-    UserDTO getUserById(Long id);
+    UserDTO getUserById(Integer id);
     UserDTO createUser(UserDTO userDTO);
-    UserDTO updateUser(Long id, UserDTO userDTO);
-    void deleteUser(Long id);
+    UserDTO updateUser(Integer id, UserDTO userDTO);
+    void deleteUser(Integer id);
 }
